fix warning C4244 on Windows
'initializing': conversion from 'uint64_t' to 'int',  possible loss of data

update optimizer_cases.c.h, too

+++ Python/optimizer_cases.c.h
@@ -1122,7 +1122,7 @@
                 ctx->done = true;
                 break;
             }
-            int returning_stacklevel = this_instr->operand1;
+            int returning_stacklevel = (int)this_instr->operand1;
             if (frame_pop(ctx, returning_code, returning_stacklevel)) {
                 break;
             }
@@ -1179,7 +1179,7 @@
                 ctx->done = true;
                 break;
             }
-            int returning_stacklevel = this_instr->operand1;
+            int returning_stacklevel = (int)this_instr->operand1;
             if (frame_pop(ctx, returning_code, returning_stacklevel)) {
                 break;
             }
@@ -3014,7 +3014,7 @@
                 ctx->done = true;
                 break;
             }
-            int returning_stacklevel = this_instr->operand1;
+            int returning_stacklevel = (int)this_instr->operand1;
             if (frame_pop(ctx, returning_code, returning_stacklevel)) {
                 break;
             }
